Fixing issues uncovered by stronger compiler warnings * Adding const qualifier for parameter 'char * fmt' -> 'const char * fmt' thereby eliminating -Wwrite-strings warnings * Adding missing cases to switch statements * Removing unused variable 'index' * Reording initialisation order of ArduinoIoTCloud in order to eliminate '-Wreorder' warnings * Adding const qualifier for parameter 'char * fmt' -> 'const char * fmt' thereby eliminating -Wwrite-strings warnings * Prefixing 'connection' with an underscore -> '_connection' in order to be consistent with all other member variables * Preventing multiple definitions of SECRET_PASS mutually overwriting one another by prefixing the correct connection type * Changing type of timeout intervals to type unsigned long - otherwise there is a type inconsistency and probable bug when checking the connection time interval * Removing duplicate ERROR case * Adding missing cases and simplifying code * Changing type of timeout intervals to type unsigned long - otherwise there is a type inconsistency and probable bug when checking the connection time interval * Using default clause for missing switch statements (no constant defined for state DISCONNECTING) * Fix -Wreorder error * Simplifying WifiConnectionManager * Replacing '#warning' with '#pragma message' * Fixing comparison between unsigned and signed type * Ignoring error stemming from warning about deprecated functions * Rectifying signed/unsigned comparison * char * arrays should be compared via strcmp * Commenting out unused parameters in order to prevent -Wunused-parameter * Removing unused variable connection buffer * Adding missing switch cases * Correct initlisation of struct tm t * Correct printf specifier for unsigned long it is %lu * Adding missing initializer for tm_isdst * Performing cast in order to guarantuee correct comparison * Fixing formatting * Disabling '-Wunused-variable' to prevent false positives

+++ src/GSMConnectionManager.h
@@ -53,17 +53,17 @@ class GSMConnectionManager : public ConnectionManager {
 
     void changeConnectionState(NetworkConnectionState _newState);
 
-    const int CHECK_INTERVAL_IDLE = 100;
-    const int CHECK_INTERVAL_INIT = 100;
-    const int CHECK_INTERVAL_CONNECTING = 500;
-    const int CHECK_INTERVAL_CONNECTED = 10000;
-    const int CHECK_INTERVAL_RETRYING = 5000;
-    const int CHECK_INTERVAL_DISCONNECTED = 1000;
-    const int CHECK_INTERVAL_ERROR = 500;
+    const unsigned long CHECK_INTERVAL_IDLE = 100;
+    const unsigned long CHECK_INTERVAL_INIT = 100;
+    const unsigned long CHECK_INTERVAL_CONNECTING = 500;
+    const unsigned long CHECK_INTERVAL_CONNECTED = 10000;
+    const unsigned long CHECK_INTERVAL_RETRYING = 5000;
+    const unsigned long CHECK_INTERVAL_DISCONNECTED = 1000;
+    const unsigned long CHECK_INTERVAL_ERROR = 500;
 
     const char *pin, *apn, *login, *pass;
     unsigned long lastConnectionTickTime;
-    int connectionTickTimeInterval;
+    unsigned long connectionTickTimeInterval;
 
 };
